+++ src/model/uniprotkb/sections/ProteinProcessingConverter.ts
@@ -1,6 +1,9 @@
-import { FeatureData } from '../../FeaturesView';
+import { FeatureData } from '../../../view/uniprotkb/components/FeaturesView';
 import FeatureTypes from '../../types/FeatureTypes';
-import { KeywordData, KeywordCategory } from '../../Keyword';
+import {
+  KeywordData,
+  KeywordCategory,
+} from '../../../view/uniprotkb/components/KeywordView';
 import { getCategoryKeywords } from '../../utils/KeywordsUtil';
 import KeywordTypes from '../../types/KeywordTypes';
