+++ src/view/uniprotkb/ExpressionSection.tsx
@@ -2,9 +2,9 @@ import React, { Fragment, FC } from 'react';
 import { Card } from 'franklin-sites';
 import { isEmpty } from '../../model/utils/utils';
 import EntrySectionType from '../../model/types/EntrySectionType';
-import { FreeText } from '../../model/FreeText';
-import { Keyword } from '../../model/Keyword';
-import XRef from '../../model/XRef';
+import { FreeText } from '../uniprotkb/components/FreeTextView';
+import { Keyword } from '../uniprotkb/components/KeywordView';
+import XRef from '../uniprotkb/components/XRefView';
 import { UniProtkbUIModel } from '../../model/uniprotkb/UniProtkbConverter';
 
 const ExpressionSection: FC<{ data: UniProtkbUIModel }> = ({ data }) => {
