+++ src/view/uniprotkb/NamesAndTaxonomySection.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment, FC } from 'react';
 import { Card } from 'franklin-sites';
-import { EntryProteinNames } from '../../model/ProteinNames';
+import { EntryProteinNames } from './components/ProteinNamesView';
 import { isEmpty } from '../../model/utils/utils';
 import EntrySectionType from '../../model/types/EntrySectionType';
 import { UniProtkbUIModel } from '../../model/uniprotkb/UniProtkbConverter';
