+++ src/view/uniprotkb/ProteinProcessingSection.tsx
@@ -2,8 +2,8 @@ import React, { Fragment, FC } from 'react';
 import { Card } from 'franklin-sites';
 import { isEmpty } from '../../model/utils/utils';
 import EntrySectionType from '../../model/types/EntrySectionType';
-import FeaturesView from '../../model/FeaturesView';
-import { Keyword } from '../../model/Keyword';
+import FeaturesView from '../uniprotkb/components/FeaturesView';
+import { Keyword } from '../uniprotkb/components/KeywordView';
 import { UniProtkbUIModel } from '../../model/uniprotkb/UniProtkbConverter';
 
 const ProteinProcessingSection: FC<{ data: UniProtkbUIModel }> = ({ data }) => {
