+++ src/view/uniprotkb/SequenceSection.tsx
@@ -2,10 +2,10 @@ import React, { Fragment, FC } from 'react';
 import { Card } from 'franklin-sites';
 import { isEmpty } from '../../model/utils/utils';
 import EntrySectionType from '../../model/types/EntrySectionType';
-import FeaturesView from '../../model/FeaturesView';
-import { Keyword } from '../../model/Keyword';
-import XRef from '../../model/XRef';
-import { SequenceViewEntry } from '../../model/SequenceView';
+import FeaturesView from '../uniprotkb/components/FeaturesView';
+import { Keyword } from '../uniprotkb/components/KeywordView';
+import XRef from '../uniprotkb/components/XRefView';
+import { SequenceViewEntry } from '../uniprotkb/components/SequenceView';
 import { UniProtkbUIModel } from '../../model/uniprotkb/UniProtkbConverter';
 
 const SequenceSection: FC<{ data: UniProtkbUIModel }> = ({ data }) => {
