+++ src/view/uniprotkb/SubcellularLocationSection.tsx
@@ -2,7 +2,7 @@ import React, { Fragment, FC } from 'react';
 import { Card } from 'franklin-sites';
 import { isEmpty } from '../../model/utils/utils';
 import EntrySectionType from '../../model/types/EntrySectionType';
-import { Keyword } from '../../model/Keyword';
+import { Keyword } from '../uniprotkb/components/KeywordView';
 import { UniProtkbUIModel } from '../../model/uniprotkb/UniProtkbConverter';
 
 const SubcellularLocationSection: FC<{ data: UniProtkbUIModel }> = ({
