+++ src/view/uniprotkb/UniProtEntryConfig.tsx
