+++ src/view/uniprotkb/components/CatalyticActivityView.tsx
@@ -1,8 +1,8 @@
 import React, { Fragment } from 'react';
 import { v1 } from 'uuid';
-import UniProtEvidenceTag from '../components/UniProtEvidenceTag';
-import { EvidenceType } from './types/modelTypes';
-import { CommentType } from './types/CommentType';
+import UniProtEvidenceTag from '../../../components/UniProtEvidenceTag';
+import { EvidenceType } from '../../../model/types/modelTypes';
+import { CommentType } from '../../../model/types/CommentType';
 
 export type CatalyticActivityData = {
   commentType: CommentType;
