+++ src/view/uniprotkb/components/GeneNamesView.tsx
