+++ src/view/uniprotkb/components/NameView.tsx
