+++ src/view/uniprotkb/components/SimpleView.tsx
